% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_plot.R
\name{bc_plot}
\alias{bc_plot}
\title{Bias Correction Plot}
\usage{
bc_plot(p, lower = 0.045, upper = 0.055, col_accept = "lightblue",
  col_reject = "indianred")
}
\arguments{
\item{p}{a vector of p-values returned from bias_correction}

\item{lower}{lower bound of what is considered an acceptable error rate}

\item{upper}{upper bound of what is considered an acceptable error rate}

\item{col_accept}{color of bias corrected p-value that falls within lower and upper}

\item{col_reject}{color of bias corrected p-value that does not fall within lower and upper}
}
\description{
This function plots the results of bias_correction. It allows for 
visuallizing error rates deemed acceptable by the user
}
\examples{
bc_plot(p, lower = 0.025, upper = 0.065, col_accept = "purple", col_reject = "orange")

}
