% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_func.R
\name{data_func}
\alias{data_func}
\title{Called by GEE Bias Correction}
\usage{
data_func(icc, v_overall, n_litters, pups_litter)
}
\description{
Called by GEE Bias Correction
}
\examples{
data_func()
}
