% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_icc.R
\name{litter_icc}
\alias{litter_icc}
\title{ICC for litter}
\usage{
litter_icc(formula, data)
}
\arguments{
\item{formula}{lme4 arugments y ~ x + (1|litter)}

\item{data}{}
}
\description{
This function computes ICC for litter. This is the residual 
variance explained by litter
}
\examples{

# specify values
b_0 <- 5
b_treat <- 2
icc <- 0.5
v_overall <- 10
n_litters <- 12
pups_litter <- 4
my_data <- data_generator(b_0 = b_0, b_treat = b_treat, 
                   icc = icc, v_overall = v_overall, 
                   n_litters = n_litters, pups_litter = pups_litter)
                   
litter_icc(y ~ treat + (1|litter), data = my_data)
}
