% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_fun.R
\name{pow_func}
\alias{pow_func}
\title{Called by prospective power
Please not this function is included so that the underlying code
can be seen. This funtion does nothing by itself, but is called
by the function prospective power analyses}
\usage{
pow_func(delta_t, icc, v_overall, n_litters, pups_litter, method,
  parameter = "treatment", correction = "d5")
}
\description{
Called by prospective power
Please not this function is included so that the underlying code
can be seen. This funtion does nothing by itself, but is called
by the function prospective power analyses
}
\examples{
power_func()
}
