% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective_power.R
\name{retrospective_power}
\alias{retrospective_power}
\title{Retrospective Power for the the treatment effect}
\usage{
retrospective_power(nsims, outcome, treatment, litter)
}
\arguments{
\item{nsims}{number of iterations. Should be > 1,000, but smaller values can build intution for expected power}

\item{outcome}{predicted observations}

\item{treatment}{(two groups: high vs low LG)}

\item{litter}{litter variable (see my_data for example)}
}
\description{
This function computes power for observed values (e.g.,
the treatment effect, icc, overall variance, etc.)
}
