% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litter_var_explained_by_treat.R
\name{litter_var_explained_by_treat}
\alias{litter_var_explained_by_treat}
\title{Between-litter variance explained by
the treatment effect}
\usage{
litter_var_explained_by_treat(outcome, treatment, litter)
}
\arguments{
\item{formula}{lme4 arugments y ~ x + (1|litter)}

\item{data}{}
}
\description{
This function computes the between_litter variance
explained by the treatment effect. Note that treatment
does not necessarily reflect and experimental design.
In this context, the treatment is a comparison between
two groups (e.g., high vs low lg)
}
\examples{

# specify values
b_0 <- 5
b_treat <- 2
icc <- 0.5
v_overall <- 10
n_litters <- 12
pups_litter <- 4
my_data <- data_generator(b_0 = b_0, b_treat = b_treat, 
                   icc = icc, v_overall = v_overall, 
                   n_litters = n_litters, pups_litter = pups_litter)
                   
                   
outcome <- my_data$y
treatment <- my_data$treatment
litter <- my_data$litter
var_explained_by_treat(outcome = outcome, 
                       treatment = treatment, l
                       itter = litter)
}
