% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective_power.R
\name{prospective_power}
\alias{prospective_power}
\title{Prospective Power Analyses}
\usage{
prospective_power(nsims, delta_t, icc, v_overall, n_litters, pups_litter,
  method, parameter = "treatment", correction)
}
\description{
This function computes power
}
