% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective_power.R
\name{prospective_power}
\alias{prospective_power}
\title{Prospective Power Analyses}
\usage{
prospective_power(nsims, delta_t, icc, v_overall, n_litters, pups_litter,
  method, parameter = "treatment", correction)
}
\arguments{
\item{nsims}{number of iterations. Should be > 1,000, but smaller values can 
build intution for expected power}

\item{delta_t}{standardized effect size. Follows Cohen's guidelines}

\item{icc}{residual variance explained by litter (0 - 0.99)}

\item{v_overall}{total variance in the model}

\item{n_litters}{number of litters (must be even)}

\item{pups_litter}{number of pups per litter (must be even)}

\item{method}{model: "MLM", "GEE", "LM" for litter means}

\item{parameter}{default is "treatment"
set to "litter" for computing
power to detect litter effect (random intercept)}

\item{correction}{bias correction for GEE. 
"d5" was used in manuscript
see saws package documentation for other options}
}
\description{
This function computes power for multilevel models, generalized estimating equations,
and litter means. For multilevel models, power for the treatment and litter effects can 
be caluculated.
}
\examples{
example that works
prospective_power(nsims =  20, delta_t = .25, icc = .5, v_overall = 20, 
                  n_litters = 50, pups_litter = 4, method = "GEE", 
                  correction = "d5", parameter = "treatment")
example that results in error (litter means cannot estimate litter effect)
prospective_power(nsims =  20, delta_t = .25, icc = .5, v_overall = 20, 
                  n_litters = 50, pups_litter = 4, method = "LM", 
                  correction = "d5", parameter = "litter")




}
