% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective_power.R
\name{prospective_power}
\alias{prospective_power}
\title{Prospective Power Analyses}
\usage{
prospective_power(nsims, delta_t, icc, v_overall, n_litters, pups_litter,
  method, parameter = "treatment", correction)
}
\description{
This function computes power
}
\examples{
example that works
prospective_power(nsims =  20, delta_t = .25, icc = .5, v_overall = 20, 
                  n_litters = 50, pups_litter = 4, method = "GEE", 
                  correction = "d5", parameter = "treatment")
example that results in error (litter means cannot estimate litter effect)
prospective_power(nsims =  20, delta_t = .25, icc = .5, v_overall = 20, 
                  n_litters = 50, pups_litter = 4, method = "LM", 
                  correction = "d5", parameter = "litter")




}
